+++ src/controllers/workareas.controllers.ts
@@ -1,19 +1,8 @@
 import { NextFunction, Request, Response } from 'express';
-import { ProjectsServices, WorkAreasServices } from '../services';
+import { PathServices, ProjectsServices, WorkAreasServices } from '../services';
 import fs from 'fs';
-// export const showWorkareas = async (
-//   req: Request,
-//   res: Response,
-//   next: NextFunction
-// ) => {
-//   try {
-//     const query = await WorkAreasServices.getAll();
-//     res.status(200).json(query);
-//   } catch (error) {
-//     next(error);
-//   }
-// };
-const path = './uploads';
+import { renameDir, setNewPath } from '../utils/fileSystem';
+
 export const showWorkArea = async (
   req: Request,
   res: Response,
@@ -51,12 +40,9 @@ export const createWorkArea = async (
 ) => {
   try {
     const { body } = req;
-    const project = await ProjectsServices.find(body.projectId);
-    const query = await WorkAreasServices.create(body, project.dir);
-    if (project && query) {
-      const newDir = query.dir + '/' + query.item + '.' + query.name;
-      fs.mkdirSync(newDir);
-    }
+    const query = await WorkAreasServices.create(body);
+    const path = await PathServices.pathArea(query.id);
+    if (query) fs.mkdirSync(path);
     res.status(201).json(query);
   } catch (error) {
     next(error);
@@ -72,14 +58,12 @@ export const updateWorkarea = async (
     const { id } = req.params;
     const { body } = req;
     const _work_area_id = parseInt(id);
-    const oldTask = await WorkAreasServices.find(_work_area_id);
+    const oldDir = await PathServices.pathArea(_work_area_id);
     const query = await WorkAreasServices.update(_work_area_id, body);
-    if (query) {
-      const oldDir = oldTask.dir + '/' + oldTask.item + '.' + oldTask.name;
-      const newDir = query.dir + '/' + query.item + '.' + query.name;
-      fs.renameSync(oldDir, newDir);
-    }
-    res.status(200).json(query);
+    const path = query.item + '.' + query.name;
+    const newDir = setNewPath(oldDir, path);
+    if (query) renameDir(oldDir, newDir);
+    res.status(200).json({ ...query, newDir });
   } catch (error) {
     next(error);
   }
@@ -93,12 +77,10 @@ export const deleteWorkarea = async (
   try {
     const { id } = req.params;
     const _work_area_id = parseInt(id);
-    const _project_id = (await WorkAreasServices.find(_work_area_id)).projectId;
-    const query = await WorkAreasServices.delete(_work_area_id, _project_id);
-    if (query) {
-      const path = query.dir + '/' + query.item + '.' + query.name;
-      fs.rmSync(path, { recursive: true });
-    }
+    const { projectId } = await WorkAreasServices.find(_work_area_id);
+    const path = await PathServices.pathArea(_work_area_id);
+    const query = await WorkAreasServices.delete(_work_area_id, projectId);
+    if (query) fs.rmSync(path, { recursive: true });
     res.status(204).json(query);
   } catch (error) {
     next(error);
